Remove 'First' section
Previously, there was a step missing instructing new users to create an
`out` directory before building. While it may have been obvious to those
that previously worked in the project, it should have existed. Now, in
upstream, the `Makefile` has been updated to perform this step
automatically, removing the need for this step to exist on the
end-user's side directly.