cross-rust: use in-tree libunwind
This builds libunwind in-tree, preventing an illegal instruction when
using libunwind built against musl.

This is a stopgap solution, as eventually we should resolve the issues
of building and linking libraries when cross compiling. As it currently
stands, everything has to be built manually when compiling - if stagex,
today, were to offer a cross-compiled librocksdb, it would likely
not be linked against the correct libc implementation, and would likely
build using musl headers instead of glibc.